Added mouse scroll NO FOLLOW

+++ src/complex.c
@@ -6,7 +6,7 @@
 /*   By: dbaladro <dbaladro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 14:01:11 by dbaladro          #+#    #+#             */
-/*   Updated: 2024/01/20 22:01:21 by dbaladro         ###   ########.fr       */
+/*   Updated: 2024/01/20 22:52:45 by madlab           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,13 @@ double	modul_sum(t_complex c)
 	return (c.reel * c.reel + c.img * c.img);
 }
 
-t_complex   init_complex(double reel, double imaginary)
+t_complex	init_complex(double reel, double imaginary)
 {
-    t_complex   c;
+	t_complex	c;
 
-    c.reel = reel;
-    c.img = imaginary;
-    return (c);
+	c.reel = reel;
+	c.img = imaginary;
+	return (c);
 }
 
 t_complex	c_add(t_complex c1, t_complex c2)
@@ -54,7 +54,7 @@ t_complex	pos_to_complex(t_pos pos, t_param param)
 	middle_line = (double) param.ref.y;
 	middle_col = (double) param.ref.x;
 	final_zoom = middle_line * param.zoom;
-	c.reel = ((double)pos.y - param.offset.y - middle_col) / final_zoom;
-	c.img = ((double)pos.x - param.offset.x - middle_line) / final_zoom;
+	c.reel = ((double)pos.y - (double)param.offset.y - middle_col) / final_zoom;
+	c.img = ((double)pos.x - (double)param.offset.x - middle_line) / final_zoom;
 	return (c);
 }
